+++ index.html
@@ -229,7 +229,7 @@
 
         <!-- Здесь будут изображения других пользователей -->
       </section>
-      <script src="js/main.js"></script>
+      <script src="js/main.js" type="module"></script>
     </main>
 
     <footer class="page-footer container">
