+++ js/data.js
@@ -1,0 +1,27 @@
+const COUNT_OBJECTS = 25;
+const DESCRIPTIONS = [
+  'Музыка',
+  'Юмор',
+  'Фото',
+  'Россия',
+  'Девушки',
+  'Лига Добра',
+  'Картинка',
+  'Кот',
+  'Ваза',
+  'Ремонт техники',
+  'Собака',
+  'Картинка с текстом',
+  'Боксёр',
+  'Потеряшка'
+];
+const CountLikes = {
+  MIN: 15,
+  MAX: 200
+};
+const CountComments = {
+  MIN: 0,
+  MAX: 200
+};
+
+export { COUNT_OBJECTS, DESCRIPTIONS, CountLikes, CountComments };
